+++ api/controllers/case-list/index.ts
@@ -6,7 +6,7 @@ import * as log4jui from '../../lib/log4jui'
 import { asyncReturnOrError } from '../../lib/util'
 import { getMutiJudCCDCases } from '../../services/ccd-store-api/ccd-store'
 import { getDecision } from '../../services/coh'
-import { getHearingByCase } from '../../services/coh-cor-api/coh-cor-api'
+import { getHearingByCase } from '../../services/cohQA'
 
 const getListTemplate = require('./templates/index')
 const { processCaseState } = require('../../lib/processors/case-state-model')
