+++ api/controllers/events/index.ts
@@ -3,7 +3,7 @@ const moment = require('moment')
 const getEventTemplate = require('./templates/index')
 const valueProcessor = require('../../lib/processors/value-processor')
 const { getCCDEvents } = require('../../services/ccd-store-api/ccd-store')
-const { getHearingIdOrCreateHearing, getOnlineHearingConversation } = require('../../services/coh-cor-api/coh-cor-api')
+import { getHearingIdOrCreateHearing, getOnlineHearingConversation } from '../../services/cohQA'
 
 function hasCOR(jurisdiction, caseType) {
     return jurisdiction === 'SSCS'
@@ -94,7 +94,7 @@ function reduceCohEvents(events) {
 }
 
 export async function getCohEvents(userId, caseId) {
-    const hearingId =  await getHearingIdOrCreateHearing(caseId)
+    const hearingId = await getHearingIdOrCreateHearing(caseId)
     const conversation = await getOnlineHearingConversation(hearingId)
     const mergedEvents = mergeCohEvents(conversation)
     return reduceCohEvents(mergedEvents)
@@ -110,7 +110,7 @@ function combineLists(lists) {
 
 function sortEvents(events) {
     return events.sort((result1, result2) =>
-    moment.duration(moment(result2.dateUtc).diff(moment(result1.dateUtc))).asMilliseconds())
+        moment.duration(moment(result2.dateUtc).diff(moment(result1.dateUtc))).asMilliseconds())
 }
 
 export async function getEvents(userId, jurisdiction, caseType, caseId) {
@@ -118,7 +118,7 @@ export async function getEvents(userId, jurisdiction, caseType, caseId) {
     const ccdEvents = await getCcdEvents(userId, jurisdiction, caseType, caseId)
 
     if (hasCOR(jurisdiction, caseType)) {
-         cohEvents = await getCohEvents(userId, caseId)
+        cohEvents = await getCohEvents(userId, caseId)
     }
 
     const combined = combineLists([ccdEvents, cohEvents])
