+++ api/services/ccdDef.ts
@@ -1,7 +1,7 @@
 import * as express from 'express'
-import { config } from '../../../config'
-import { http } from '../../lib/http'
-import { getHealth, getInfo } from '../../lib/util'
+import { config } from '../../config'
+import { http } from '../lib/http'
+import { getHealth, getInfo } from '../lib/util'
 
 const url = config.services.ccd_def_api
